Update GenericGroovyTemplateResolver.groovy
The plugin currently fails 2 test cases on Windows because of path separator differences. The problem also manifests in Grails 4.0.4 since it updated to 2.0.2 of the views plugin, on Windows Grails controllers fail to resolve GSON views and either throw an exception or fall back to a default JSON renderer.

I haven't dug into it too far, but it appears that resolveTemplateName is called in a couple of contexts, sometimes with a filesystem path and sometimes with a resource path, so I think the right thing to do is remove the filesystem or resource path separator.

I've verified that the proposed change passes the tests when built on Windows, and also fixes the GSON view resolution problem in Grails 4.0.4 on Windows.