build: Update Groovy Joint workflow
The groovy version was not resolved correctly previously. With this change it gets set as three different environment vars. CI_GROOVY_VERSION and GROOVY_VERSION as regular environment vars and with ORG_GRADLE_PROJECT_groovyVersion that sets the groovyVersion project property. This allows for some choices in how to handle this in the build script.

This change also adds a section to the settings.gradle file to allow for overriding the groovy version in the version catalog with the GROOVY_VERSION env variable.