chore(build): Modernise and clean up build (#504)
* chore(build): Modernise and clean up build

- Implement buildSrc to isolate build related code and deps
- Rename gradle plugin project dir to gradle-plugin
- Add gradle dir for version catalogs
- Implement version catalogs to handle versions
- Remove redundant gradle files from subprojects
- Move grailsCentralPublishing.gradle to gradle dir
- Use project.layout to reference files and dirs
- Make build more type-safe with IDE code completion
- Unify quote usage
- Clean up dependencies and tighten their scopes
- Use lazy configuration where possible
- Remove unused ext properties
- Create workarounds for grails-gradle-plugin exposing groovy
- Remove used resources and attribute to asciidoctor task
- Fix classpath issues with groovydoc task
- Change 'docs' task as there is nothing to copy
- Change jar manifest to reflect projects properties
- Remove redundant Application.groovy from views-markup
- Update example projects
- Add missing CompileStatic and make compatible with Groovy 3.0.21-SNAPSHOT

* fix(build): Move Snapshot repos to the right place

The snapshot repositories had accidentally been placed outside
the configuration of the PublishingExtension.

* build(chore): Clean up secret passing from GH action

By using the ORG_GRADLE_PROJECT_** env vars build files can
be cleaned up from checking at multiple places for the values.

* ci: Replace Java 14 with Java 17 in build matrix

* build: Declare grailsVersion in one place

With this change the Grails version is only declared once.

* build: Update Groovy Joint workflow

The groovy version was not resolved correctly previously. With this change it gets set as three different environment vars. CI_GROOVY_VERSION and GROOVY_VERSION as regular environment vars and with ORG_GRADLE_PROJECT_groovyVersion that sets the groovyVersion project property. This allows for some choices in how to handle this in the build script.

This change also adds a section to the settings.gradle file to allow for overriding the groovy version in the version catalog with the GROOVY_VERSION env variable.

* build: Create configuration for Groovydoc

The Groovydoc Gradle task need to be run with the same groovy
version as Gradle.

* build: Change workflow JDK to temurin

AdoptOpenJDK got moved to Eclipe Temurin and won't
be updated anymore.

* build: Cleanup handling of sonatype credentials

With this change there is no need to check both environment
variables and project properties.
Also move the sonatype repo info to a project property.

* chore(deps): Update grails-gradle-plugin to 6.1.2

* chore(deps): Update grails to 6.1.2

* build: Make groovy workflow version independent

With this change the workflow will use the versions of groovy
and gradle enterprise that are defined in the build rather
than hard-coding them in the actions file.

* build: Minor cleanups

* ci: Remove publishing of test reports

These can now be viewed in build scans.

* build: Remove unused import

* ci: Remove java version matrix for release

No need to run this on multiple versions

* build: Change author and git email

Puneet has a new email